changed name from example to palone, code style reformat
