+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.PlanAhead"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.palone.planahead.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.PlanAhead">
