--- app/src/main/java/com/example/planahead/data/HomeScreenUIState.kt
@@ -1,6 +1,0 @@
-package com.example.planahead.data
-
-import com.example.planahead.data.task.Task
-import kotlinx.coroutines.flow.Flow
-
-data class HomeScreenUIState(var isLoading:Boolean = false)
