--- app/src/main/java/com/example/planahead/data/ScreenProperties.kt
@@ -1,5 +1,0 @@
-package com.example.planahead.data
-
-sealed class ScreensProperties(val route: String, val name: String, val selected: Boolean = false) {
-    object HomeScreen : ScreensProperties("home_screen", "Home Screen", true)
-}
