--- app/src/main/java/com/example/planahead/data/task/properties/AlertType.kt
@@ -1,7 +1,0 @@
-package com.example.planahead.data.task.properties
-
-object AlertType {
-    val ALARM = "ALARM" // works as alarm clock
-    val NOTIFICATION = "NOTIFICATION"  // just a system notification
-    val PERSISTENT_NOTIFICATION = "PERSISTENT_NOTIFICATION" // a system notification that you can't get rid of
-}
