+++ app/src/main/java/com/palone/planahead/data/HomeScreenUIState.kt
@@ -1,0 +1,4 @@
+package com.palone.planahead.data
+
+
+data class HomeScreenUIState(var isLoading: Boolean = false)
