+++ app/src/main/java/com/palone/planahead/data/database/TaskDao.kt
@@ -1,0 +1,20 @@
+package com.palone.planahead.data.database
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Query
+import androidx.room.Upsert
+import com.palone.planahead.data.task.Task
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface TaskDao {
+    @Upsert
+    suspend fun upsertTask(task: Task)
+
+    @Delete
+    suspend fun deleteTask(task: Task)
+
+    @Query("SELECT * FROM task ORDER BY id DESC")
+    fun getAllTasks(): Flow<List<Task>>
+}
