+++ app/src/main/java/com/palone/planahead/data/database/TaskRepository.kt
@@ -1,0 +1,15 @@
+package com.palone.planahead.data.database
+
+import com.palone.planahead.data.task.Task
+import kotlinx.coroutines.flow.Flow
+
+class TaskRepository(private val taskDao: TaskDao) {
+    val allTasks: Flow<List<Task>> = taskDao.getAllTasks()
+    suspend fun upsert(task: Task) {
+        taskDao.upsertTask(task)
+    }
+
+    suspend fun delete(task: Task) {
+        taskDao.deleteTask(task)
+    }
+}
