+++ app/src/main/java/com/palone/planahead/data/task/properties/AlertTrigger.kt
@@ -1,0 +1,6 @@
+package com.palone.planahead.data.task.properties
+
+object AlertTrigger {
+    val LOCATION = "LOCATION"// trigger an alert by the device's location
+    val TIME = "TIME"// trigger an alert by the device's time
+}
