+++ app/src/main/java/com/palone/planahead/data/task/properties/TaskType.kt
@@ -1,0 +1,8 @@
+package com.palone.planahead.data.task.properties
+
+object TaskType {
+    val ONE_TIME = "ONE_TIME" // one time event
+    val CRON = "CRON"  // repeated event by schedule
+    val CHORE = "CHORE"  // repeated event by time
+}
+
